requirements.txt
PySocks
requests
python-telegram-bot
sqlalchemy
